CAN_Control:
    Thêm mode NODE_REQ_SPEED_ANGLE
CAN_FuncHandle:
    CAN_SpeedBLDC_AngleDC đổi float bldcSpeed thành int bldcSpeed và gửi lên số xung encoder X4 đo được của NodeSwerve_F103, lý do:
        Khi dùng RTR thì tốc độ BLDC được PID tính toán của NodeSwerve_F103 bị ảnh hưởng (dao động nhiều hơn với Setpoint so với không có RTR) đồng thời khi gửi giá trị này để Main_F407 nhận được thì số dao động nhiều đến mức không thể đọc được tốc độ BLDC
        Tuy nhiên với góc xoay không có sự thay đổi giá trị nhiều thì giá trị góc quay (gửi kèm với giá trị tốc độ BLDC) không bị ảnh hường và đọc góc vẫn chính xác
        Với số xung encoder đọc được thì giá trị giữa NodeSwerve_F103 và Main_F407 hoàn toàn giống nhau
Bộ lọc Main_F407 không còn CANCTRL_MODE_MOTOR_SPEED_ANGLE mà thay thành CANCTRL_MODE_NODE_REQ_SPEED_ANGLE

Hiện tại việc request số xung đọc về và góc quay còn gặp vấn đề sau (yêu cầu là lấy nhanh nhất số xung và góc quay của các cụm bánh -> hạn chế sử dụng delay mỗi lần request):
    giữa các khoảng request RTR cần có một khoảng nghỉ delay khoảng osDelay(1) hoặc loop lệnh __NOP() khoảng 8000(đối với CAN baud là 500kbit) hoặc 3000(với CAN baud là 1Mbit)

+++ Main_F407/CAN/Inc/CAN_FuncHandle.h
@@ -11,7 +11,7 @@
 #include "BoardParameter.h"
 
 typedef struct CAN_SpeedBLDC_AngleDC{
-	float bldcSpeed;
+	int bldcSpeed;
 	float dcAngle;
 }CAN_SpeedBLDC_AngleDC;
